+++ salary_calc/CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+## 1.2.1 - 2020-11-15
+### Changed
+- 부양가족, 20세이하 자녀 설정범위 초과시 오류메시지 문구 수정
+- 넘패드 높이 약간 축소
+- IntStepper double지원을 위해 NumStepper로 변경 및 flutter_jk로 이동
+- KrUtils 삭제 후 flutter_jk 참조하도록 변경
+- firebase-analytics 17.2.2 -> 17.2.3 변경
 
 ## 1.2.0 - 2020-09
 ### Changed